feat. 리뷰반영: bank의 역할을 winningTicket이 주체가 될 수 있도록 수정하였습니다.
